Add files via upload
The purpose of this Capstone project was to use Supervised Learning methods to predict the 30-day readmission rate for Heart Failure at hospitals across the United States. The importance of features used in the model can be harnessed to assist hospital administration and clinical staff in an effort to reduce 30-day readmissions.  Data from The Centers for Medicare & Medicaid Services (CMS) "Hospital Compare" website was used for this project.