Add files via upload
Using the National Health and Nutrition Examination Survey (NHANES) to Predict Heart Disease:
According to the Centers for Disease Control and Prevention (CDC), heart disease is the leading cause of death for men, women, and most racial and ethnic groups in the United States. Although the link between diet/nutrition and heart disease has long been established, this link is incompletely understood. 
The NHANES, a major program of the National Center for Health Statistics (NCHS), part of the CDC, has examined a nationally representative sample of 5,000 people each year since 1999. The collected data contains not only detailed information on nutrient intake and diet behaviors, it also includes laboratory tests and physiological measurements, providing a wealth of data points. 
The purpose of this project is to use the NHANES data in an attempt to better understand the link between heart disease and nutrition in predicting heart disease. Biostatistics was used to further explore the relationship between heart disease and age/gender.